INDY-1645: Fix for config dependencies bug
