+++ plenum/test/monitoring/test_throughput_median_avg.py
@@ -3,7 +3,8 @@ import functools
 import pytest
 
 from plenum.server.instances import Instances
-from plenum.server.monitor import Monitor, MedianLowStrategy
+from plenum.server.monitor import Monitor
+from plenum.common.strategies import MedianLowStrategy
 from plenum.test.testing_utils import FakeSomething
 
 NUM_OF_REPLICAS = 5
